# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ti Saúde API",
        "description": "Documentação da API do desafio fullstack",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/planos-saude": {
            "get": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Lista todos os planos de saúde",
                "operationId": "08c06c7b4a3c6d45c117c106b4b68243",
                "responses": {
                    "200": {
                        "description": "Planos listados com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Cria um novo plano de saúde",
                "operationId": "8b05c7e6b326b8bab04af9e6c3e692f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plano_descricao"
                                ],
                                "properties": {
                                    "plano_codigo": {
                                        "type": "string",
                                        "example": "UNIMED"
                                    },
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Unimed Nacional"
                                    },
                                    "plano_telefone": {
                                        "type": "string",
                                        "example": "08000001234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plano criado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/planos-saude/{id}": {
            "get": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Busca um plano de saúde por ID",
                "operationId": "5aba129ef993bbf346bb0d8e02819b5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano encontrado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Plano não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Atualiza um plano de saúde",
                "operationId": "0c0b0b451be86fcf5d20d2008e45fc5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plano_descricao"
                                ],
                                "properties": {
                                    "plano_codigo": {
                                        "type": "string",
                                        "example": "UNIMED"
                                    },
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Unimed Nacional"
                                    },
                                    "plano_telefone": {
                                        "type": "string",
                                        "example": "08000001234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plano atualizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Plano não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Remove um plano de saúde",
                "operationId": "5eca697e12141f13a47b71b7490a1317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano removido com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Plano não encontrado"
                    },
                    "409": {
                        "description": "Plano possui vínculos com pacientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}